refactor(session): extract round bridge into sessionRoundBridge.js (-49% LOC)
Token optimization A2: estrae le 600+ LOC del round-model bridge
(guard, adapter, placeholderResolveAction, handleLegacyAttackViaRound,
handleTurnEndViaRound, 4 round endpoints) in un modulo separato con
factory pattern createRoundBridge(deps).

session.js: 1665 → 997 LOC (-40% da PR precedente, -49% da originale 1967)

Totale split:
  sessionConstants.js:   58 LOC (costanti + isRoundModelEnabled)
  sessionHelpers.js:    248 LOC (15 pure functions)
  sessionRoundBridge.js: 602 LOC (round bridge factory + endpoints)
  session.js:           997 LOC (createSessionRouter closure only)
  TOTAL:              1905 LOC (was 1967 — no code lost, 62 LOC overhead)

createRoundBridge factory pattern: prende deps da closure session.js
(performAttack, buildAttackEvent, buildMoveEvent, emitKillAndAssists,
appendEvent, persistEvents, consumeCapPt, declareSistemaIntents,
roundOrchestrator, rng, resolveSession) e ritorna:
- roundModelGuard, adaptSessionToRoundState, ensureRoundState
- placeholderResolveAction
- handleLegacyAttackViaRound, handleTurnEndViaRound
- mountRoundEndpoints(router) — mounts 4 round routes

session.js ora chiama roundBridge.mountRoundEndpoints(router) per
registrare i 4 endpoint round-based. Le funzioni wrapper
handleLegacyAttackViaRound/handleTurnEndViaRound sono destructured
dal bridge e usate inline nei handler /action e /turn/end.

Token impact: quando Claude edita session.js (legacy combat logic),
legge 997 LOC. Round bridge logic (602 LOC) non caricata. Savings
atteso: ~30% per combat edit session vs pre-split.

Validazione:
- node --test tests/api/*.test.js tests/services/*.test.js tests/ai/*.test.js
  -> 254/254 verdi
- prettier -> verde
- governance strict -> 0 errors

Rollback: git revert rimette tutto inline in session.js (~1665 LOC).
