+++ apps/backend/routes/session.js
@@ -85,6 +85,7 @@ const {
   isBackstab,
   facingFromMove,
 } = require('./sessionHelpers');
+const { createRoundBridge } = require('./sessionRoundBridge');
 
 // ADR-2026-04-16: round-based combat model migration. PR 2 di N —
 // endpoint stubs dietro feature flag USE_ROUND_MODEL. Il modulo e'
@@ -442,6 +443,31 @@ function createSessionRouter(options = {}) {
     manhattanDistance,
     gridSize: GRID_SIZE,
   });
+
+  // Round bridge factory — all round-model functions live in sessionRoundBridge.js.
+  // Orchestrator created here (needs rng from closure) with no-op resolveAction;
+  // the bridge injects real resolveAction per-request via resolveRoundPure.
+  const roundOrchestrator = createRoundOrchestrator({
+    resolveAction: (state) => ({
+      nextState: JSON.parse(JSON.stringify(state)),
+      turnLogEntry: { damage_applied: 0, healing_applied: 0 },
+    }),
+    defaultRng: rng,
+  });
+  const roundBridge = createRoundBridge({
+    performAttack,
+    buildAttackEvent,
+    buildMoveEvent,
+    emitKillAndAssists,
+    appendEvent,
+    persistEvents,
+    consumeCapPt,
+    declareSistemaIntents,
+    roundOrchestrator,
+    rng,
+    resolveSession,
+  });
+  const { handleLegacyAttackViaRound, handleTurnEndViaRound } = roundBridge;
 
   // SPRINT_020: helper riutilizzabile che avanza attraverso tutti i turni
   // IA non-player fino a fermarsi su un player vivo (o nessuno). Ritorna
@@ -954,701 +980,9 @@ function createSessionRouter(options = {}) {
   // Le 4 route qui sotto abilitano il nuovo modello round-based
   // (shared-planning → commit → resolve) descritto in
   // ADR-2026-04-16-session-engine-round-migration.md e implementato
-  // in apps/backend/services/roundOrchestrator.js (port in JS del
-  // Python orchestrator).
-  //
-  // Feature flag: `USE_ROUND_MODEL`. Se falsa, ogni endpoint ritorna
-  // 503. Se vera, il session object accumula uno state parallelo
-  // `session.roundState` in shape nativa dell'orchestrator
-  // (pending_intents, round_phase, units con hp/ap/reactions dict).
-  //
-  // Il flusso legacy (/start, /action, /turn/end, /end, /state, /:id/vc)
-  // e' intatto: la mappa `sessions` e le funzioni esistenti non sono
-  // toccate. Le due superfici coesistono. La migrazione del flusso
-  // legacy ai round endpoints (wrapper + refactor AI) e' scope di PR
-  // successive.
-  //
-  // Il `resolveAction` iniettato e' un **placeholder minimale** per
-  // PR 2: gestisce `attack` applicando 3 HP di danno fisso al target,
-  // `heal` con healing deterministico, e gli altri type come no-op
-  // con consumo AP. Il wiring al resolver reale (performAttack +
-  // traitEffects) e' scope di PR 4 (legacy wrappers).
 
-  function roundModelGuard(_req, res) {
-    if (!isRoundModelEnabled()) {
-      res.status(503).json({
-        error: 'round_model_disabled',
-        message:
-          'Feature flag USE_ROUND_MODEL non attivo. Imposta USE_ROUND_MODEL=true per abilitare gli endpoint round-based (vedi ADR-2026-04-16).',
-      });
-      return false;
-    }
-    return true;
-  }
-
-  /**
-   * Adapter session.js units -> orchestrator units.
-   *
-   * La session in-memory usa `hp`/`max_hp` scalari, `ap`/`ap_remaining`,
-   * `status` come oggetto, `guardia` come parry budget. L'orchestrator
-   * vuole `hp: {current, max}`, `ap: {current, max}`,
-   * `reactions: {current, max}`, `statuses: []`.
-   *
-   * Il mapping e' idempotente: chiamarlo piu' volte sullo stesso
-   * session object ritorna sempre la stessa shape. Preserva `tier`
-   * (default 1) e `stress` (default 0) per i predicates DSL.
-   */
-  function adaptSessionToRoundState(session) {
-    const units = (session.units || []).map((u) => {
-      const statusObj = u.status || {};
-      const statuses = [];
-      for (const [id, turns] of Object.entries(statusObj)) {
-        if (Number(turns) > 0) {
-          statuses.push({
-            id,
-            intensity: 1,
-            remaining_turns: Number(turns),
-          });
-        }
-      }
-      return {
-        id: String(u.id),
-        hp: {
-          current: Number(u.hp || 0),
-          max: Number(u.max_hp || u.hp || 0),
-        },
-        ap: {
-          current: Number(u.ap_remaining != null ? u.ap_remaining : u.ap || 0),
-          max: Number(u.ap || 0),
-        },
-        reactions: { current: 1, max: 1 },
-        initiative: Number(u.initiative || 0),
-        tier: Number(u.tier || 1),
-        stress: Number(u.stress || 0),
-        statuses,
-        reaction_cooldown_remaining: Number(u.reaction_cooldown_remaining || 0),
-      };
-    });
-    return {
-      session_id: session.session_id,
-      turn: Number(session.turn || 1),
-      round_phase: null,
-      pending_intents: [],
-      units,
-      log: [],
-    };
-  }
-
-  /**
-   * Ensure `session.roundState` e' inizializzato. Chiamato lazy
-   * all'ingresso dei 4 endpoint round-based. Se non esiste, lo
-   * costruisce dall'adattamento delle units legacy.
-   */
-  function ensureRoundState(session) {
-    if (!session.roundState) {
-      session.roundState = adaptSessionToRoundState(session);
-    }
-    return session.roundState;
-  }
-
-  /**
-   * Placeholder resolveAction per PR 2. Gestisce:
-   *   - attack: infligge 3 HP di danno fisso al target, consuma AP
-   *   - heal: applica 3 HP di healing clampato a max_hp, consuma AP
-   *   - defend/parry/ability/move: consuma AP, nessun effect
-   *
-   * Restituisce `{ nextState, turnLogEntry }` in shape orchestrator
-   * (turn_log_entry con damage_applied / healing_applied).
-   *
-   * PR 4 scope: replace con wiring al `performAttack` reale di
-   * session.js (con trait effects, fairness cap, status system, ...).
-   */
-  function placeholderResolveAction(state, action, _catalog, _rng) {
-    const next = JSON.parse(JSON.stringify(state));
-    const actorId = String(action.actor_id || '');
-    const targetId = action.target_id ? String(action.target_id) : null;
-    const actor = next.units.find((u) => u.id === actorId);
-    if (actor && actor.ap) {
-      actor.ap.current = Math.max(0, Number(actor.ap.current || 0) - Number(action.ap_cost || 0));
-    }
-    let damageApplied = 0;
-    let healingApplied = 0;
-    if (action.type === 'attack' && targetId) {
-      const target = next.units.find((u) => u.id === targetId);
-      if (target && target.hp) {
-        damageApplied = 3;
-        target.hp.current = Math.max(0, Number(target.hp.current || 0) - damageApplied);
-      }
-    } else if (action.type === 'heal' && targetId) {
-      const target = next.units.find((u) => u.id === targetId);
-      if (target && target.hp) {
-        const missing = Number(target.hp.max || 0) - Number(target.hp.current || 0);
-        healingApplied = Math.max(0, Math.min(3, missing));
-        target.hp.current = Number(target.hp.current || 0) + healingApplied;
-      }
-    }
-    const turnLogEntry = {
-      turn: Number(next.turn || 1),
-      action: { ...action },
-      damage_applied: damageApplied,
-      healing_applied: healingApplied,
-    };
-    (next.log = next.log || []).push(turnLogEntry);
-    return { nextState: next, turnLogEntry };
-  }
-
-  // Orchestrator con closure-scoped deps. Gli unit test del modulo
-  // (tests/services/roundOrchestrator.test.js) validano la purezza
-  // delle funzioni — qui ci limitiamo a fornire resolveAction +
-  // rng = Math.random. Questo orchestrator usa placeholderResolveAction
-  // per i 4 endpoint stub /declare-intent, /clear-intent, /commit-round,
-  // /resolve-round. Il wrapper legacy /action (PR 4) usa invece una
-  // factory separata `createLegacyAttackViaRound` che binda
-  // `performAttack` reale al session specifico.
-  const roundOrchestrator = createRoundOrchestrator({
-    resolveAction: placeholderResolveAction,
-    defaultRng: rng,
-  });
-
-  /**
-   * PR 4: wiring wrapper legacy /action flag-on.
-   *
-   * Prende un session object (legacy shape) + action body body
-   * { actor_id, target_id, action_type } e lo esegue attraverso il
-   * round flow (planning → commit → resolve) con un resolveAction
-   * bindato al vero `performAttack` della closure session.js.
-   *
-   * Scope PR 4: solo action_type='attack'. Per altri type il caller
-   * deve cadere sul flusso legacy (il wrapper ritorna null).
-   *
-   * Ritorna un oggetto legacy-shape (roll/mos/result/pt/damage_dealt/
-   * target_hp/trait_effects/actor_position/target_position/parry)
-   * pronto per `res.json(...)`. Il wrapper aggiorna session.roundState
-   * per mantenere il dual-state coerente con gli altri endpoint
-   * round-based.
-   *
-   * Note:
-   * - Il performAttack muta session.units in place; dopo la chiamata
-   *   ri-adattiamo session.units nello state orchestrator per
-   *   mantenere hp/ap sincronizzati nel dual-state.
-   * - Cap PT e validazioni legacy (range, AP) sono eseguite PRIMA
-   *   di entrare nel round flow (coerenza con la path legacy).
-   * - Il log event e il kill/assist emission restano invariati.
-   */
-  async function handleLegacyAttackViaRound({ session, actor, target, requestedCapPt }) {
-    // Costruisci synthetic round action shape per l'orchestrator.
-    const roundAction = {
-      id: `legacy-attack-${actor.id}-${session.action_counter}`,
-      type: 'attack',
-      actor_id: actor.id,
-      target_id: target.id,
-      ability_id: null,
-      ap_cost: 1,
-      channel: null,
-      damage_dice: { count: 1, sides: 6, modifier: 2 },
-    };
-
-    // Bind resolveAction al session specifico. Il wrapper interno
-    // chiama performAttack (che muta session.units), poi sincronizza
-    // la mutazione con lo state orchestrator.
-    const capturedResults = {
-      result: null,
-      evaluation: null,
-      damageDealt: 0,
-      killOccurred: false,
-      parry: null,
-    };
-    const realResolveAction = (state, action, _catalog, _rng) => {
-      // Deep clone dello state orchestrator (shape {units: [{hp:{current,max}...}]})
-      const next = JSON.parse(JSON.stringify(state));
-      if (action.type === 'attack' && action.target_id) {
-        // performAttack muta session.units (legacy shape) in place
-        const res = performAttack(session, actor, target);
-        capturedResults.result = res.result;
-        capturedResults.evaluation = res.evaluation;
-        capturedResults.damageDealt = res.damageDealt;
-        capturedResults.killOccurred = res.killOccurred;
-        capturedResults.parry = res.parry;
-        // Sincronizza hp/ap da session.units → state orchestrator units
-        for (const uOrch of next.units) {
-          const uLegacy = session.units.find((u) => u.id === uOrch.id);
-          if (uLegacy) {
-            if (uOrch.hp) {
-              uOrch.hp.current = Number(uLegacy.hp || 0);
-              uOrch.hp.max = Number(uLegacy.max_hp || uLegacy.hp || 0);
-            }
-            if (uOrch.ap) {
-              uOrch.ap.current = Number(
-                uLegacy.ap_remaining != null ? uLegacy.ap_remaining : uLegacy.ap || 0,
-              );
-              uOrch.ap.max = Number(uLegacy.ap || 0);
-            }
-          }
-        }
-        const turnLogEntry = {
-          turn: Number(next.turn || 1),
-          action: { ...action },
-          damage_applied: Number(res.damageDealt || 0),
-          roll: res.result.roll,
-          mos: res.result.mos,
-          pt_gained: res.result.pt,
-          hit: res.result.hit,
-          die: res.result.die,
-        };
-        (next.log = next.log || []).push(turnLogEntry);
-        return { nextState: next, turnLogEntry };
-      }
-      // Fallback placeholder (non dovrebbe mai trigger per PR 4 scope)
-      return placeholderResolveAction(state, action, _catalog, _rng);
-    };
-
-    // Consuma AP legacy PRIMA di entrare nel round flow (parita' con
-    // il path legacy che decrementa ap_remaining inline).
-    actor.ap_remaining = Math.max(0, (actor.ap_remaining ?? actor.ap) - 1);
-    const hpBefore = target.hp;
-    const targetPositionAtAttack = { ...target.position };
-
-    // Costruisci round state fresco da session.units + fai round cycle
-    // completo (declare → commit → resolve) per esercitare la phase
-    // machine anche in wrapper single-actor.
-    session.roundState = adaptSessionToRoundState(session);
-    let cur = session.roundState;
-    if (cur.round_phase !== PHASE_PLANNING) {
-      cur = roundOrchestrator.beginRound(cur).nextState;
-    }
-    cur = roundOrchestrator.declareIntent(cur, actor.id, roundAction).nextState;
-    cur = roundOrchestrator.commitRound(cur).nextState;
-    // Usiamo l'API pure resolveRound del modulo, non l'orchestrator
-    // factory (che e' bindato al placeholder). Richiediamo resolveRound
-    // esportato dal modulo.
-    const result = resolveRoundPure(cur, null, rng, realResolveAction);
-    session.roundState = result.nextState;
-
-    // Emetti event legacy come il path originale
-    const event = buildAttackEvent({
-      session,
-      actor,
-      target,
-      result: capturedResults.result,
-      evaluation: capturedResults.evaluation,
-      damageDealt: capturedResults.damageDealt,
-      hpBefore,
-      targetPositionAtAttack,
-    });
-    if (capturedResults.parry) event.parry = capturedResults.parry;
-    if (requestedCapPt > 0) {
-      event.cost = { cap_pt: requestedCapPt };
-      consumeCapPt(session, requestedCapPt);
-    }
-    await appendEvent(session, event);
-    if (capturedResults.killOccurred) {
-      await emitKillAndAssists(session, actor, target, event);
-    }
-
-    return {
-      roll: capturedResults.result.roll,
-      mos: capturedResults.result.mos,
-      result: capturedResults.result.hit ? 'hit' : 'miss',
-      pt: capturedResults.result.pt,
-      damage_dealt: capturedResults.damageDealt,
-      target_hp: target.hp,
-      trait_effects: capturedResults.evaluation.trait_effects,
-      cap_pt_used: session.cap_pt_used,
-      cap_pt_max: session.cap_pt_max,
-      actor_position: { ...actor.position },
-      target_position: targetPositionAtAttack,
-      parry: capturedResults.parry,
-      state: publicSessionView(session),
-      // Round flow metadata (only present when flag on)
-      round_wrapper: true,
-      round_phase: result.nextState.round_phase,
-    };
-  }
-
-  /**
-   * PR 5 (ADR-2026-04-16 M5b): wiring wrapper legacy /turn/end flag-on.
-   *
-   * Quando USE_ROUND_MODEL=true, il /turn/end endpoint delega alla
-   * round-based pipeline:
-   *   1. Build roundState fresco da session.units
-   *   2. beginRound (refresh AP/reactions, decay status, bleeding tick
-   *      su tutte le unita')
-   *   3. declareSistemaIntents(session) per tutti i SIS
-   *   4. Per ogni intent -> declareIntent(roundState, unit_id, action)
-   *   5. commitRound
-   *   6. resolveRound con real resolveAction che:
-   *      - attack: delega a performAttack (esistente)
-   *      - move: aggiorna session.units.position + append move event
-   *      - altri type: consuma solo AP
-   *   7. Sync session.units con roundState post-resolve
-   *   8. session.turn += 1 (per compat UI legacy)
-   *
-   * Ritorna: { turn, active_unit, ia_actions, side_effects, state,
-   *            round_wrapper, round_phase }.
-   *
-   * side_effects derivato da beginRound expired+bleeding (shape
-   * simile a quello del path legacy: { unit_id, damage, hp_after, killed }).
-   *
-   * ia_actions derivato dai turn_log_entries del round: shape
-   * compatibile con il legacy (actor='sistema', type, target, ecc).
-   */
-  async function handleTurnEndViaRound(session) {
-    // 1. Build fresh roundState from session.units. Include player +
-    //    SIS units. Il risultato ha round_phase null (verra' settato
-    //    a 'planning' da beginRound).
-    session.roundState = adaptSessionToRoundState(session);
-
-    // 2. beginRound: refresh AP/reactions per ogni unit + bleeding tick +
-    //    decay status + reaction cooldown decrement. Ritorna bleedingTotal
-    //    aggregato ma non ha granularita' per-unit come il path legacy.
-    //    Per preservare side_effects shape, calcoliamo noi il tick HP qui.
-    const bleedingEvents = [];
-    for (const unit of session.units) {
-      if (!unit || !unit.status || Number(unit.hp || 0) <= 0) continue;
-      const bleedTurns = Number(unit.status.bleeding) || 0;
-      if (bleedTurns <= 0) continue;
-      const dmg = 1;
-      const hpBefore = unit.hp;
-      unit.hp = Math.max(0, unit.hp - dmg);
-      session.damage_taken[unit.id] = (session.damage_taken[unit.id] || 0) + dmg;
-      await appendEvent(session, {
-        ts: new Date().toISOString(),
-        session_id: session.session_id,
-        action_type: 'bleeding',
-        actor_id: unit.id,
-        actor_species: unit.species,
-        actor_job: unit.job,
-        target_id: unit.id,
-        turn: session.turn,
-        damage_dealt: dmg,
-        result: 'hit',
-        hp_before: hpBefore,
-        hp_after: unit.hp,
-        bleeding_remaining: bleedTurns - 1,
-        trait_effects: [],
-      });
-      bleedingEvents.push({
-        unit_id: unit.id,
-        damage: dmg,
-        hp_after: unit.hp,
-        killed: unit.hp === 0,
-      });
-    }
-    // Reset AP + decrement status per tutte le unita' (equivalente
-    // end-of-round). Il legacy path fa questo solo sulla corrente prima
-    // di advance, ma nel round model il refresh e' globale.
-    for (const unit of session.units) {
-      if (!unit) continue;
-      const fractureActive = Number(unit.status?.fracture) > 0;
-      unit.ap_remaining = fractureActive ? Math.min(1, unit.ap) : unit.ap;
-      if (unit.status) {
-        for (const key of Object.keys(unit.status)) {
-          const v = Number(unit.status[key]);
-          if (v > 0) unit.status[key] = v - 1;
-        }
-      }
-    }
-    session.roundState = adaptSessionToRoundState(session);
-    session.roundState = roundOrchestrator.beginRound(session.roundState).nextState;
-
-    // 3. declareSistemaIntents: emette intents per tutti i SIS vivi
-    const { intents, decisions } = declareSistemaIntents(session);
-
-    // 4. Declare ogni intent nel roundState
-    let cur = session.roundState;
-    for (const { unit_id, action } of intents) {
-      cur = roundOrchestrator.declareIntent(cur, unit_id, action).nextState;
-    }
-    session.roundState = cur;
-
-    // 5. commitRound
-    session.roundState = roundOrchestrator.commitRound(session.roundState).nextState;
-
-    // 6. Real resolveAction: gestisce attack (performAttack) + move
-    //    (position update + appendEvent). Capture turn_log entries per
-    //    costruire ia_actions legacy-shape.
-    const iaActions = [];
-    const realResolveAction = (state, action, _catalog, _rng) => {
-      const next = JSON.parse(JSON.stringify(state));
-      const actorId = String(action.actor_id || '');
-      const actor = session.units.find((u) => u.id === actorId);
-      if (!actor) {
-        return { nextState: next, turnLogEntry: { action, skipped: 'no_actor' } };
-      }
-      let turnLogEntry = {
-        turn: Number(next.turn || 1),
-        action: { ...action },
-        damage_applied: 0,
-      };
-
-      if (action.type === 'attack' && action.target_id) {
-        const target = session.units.find((u) => u.id === String(action.target_id));
-        if (!target || target.hp <= 0) {
-          turnLogEntry.skipped = 'target_dead';
-          iaActions.push({
-            actor: 'sistema',
-            unit_id: actorId,
-            type: 'skip',
-            reason: 'target_dead',
-            ia_rule: action.source_ia_rule,
-          });
-        } else {
-          const hpBefore = target.hp;
-          const targetPosAtk = { ...target.position };
-          const res = performAttack(session, actor, target);
-          actor.ap_remaining = Math.max(0, (actor.ap_remaining ?? actor.ap) - 1);
-          const event = buildAttackEvent({
-            session,
-            actor,
-            target,
-            result: res.result,
-            evaluation: res.evaluation,
-            damageDealt: res.damageDealt,
-            hpBefore,
-            targetPositionAtAttack: targetPosAtk,
-          });
-          event.actor_id = 'sistema';
-          event.actor_species = actor.species;
-          event.actor_job = actor.job;
-          event.ia_rule = action.source_ia_rule;
-          event.ia_controlled_unit = actor.id;
-          if (res.parry) event.parry = res.parry;
-          // appendEvent e' async ma qui nel resolveAction siamo in ctx sync.
-          // Usiamo la versione non-async (push + persist async deferred).
-          session.events.push(event);
-          session.action_counter++;
-          if (res.killOccurred) {
-            // Assists are tracked via emitKillAndAssists ma e' async.
-            // Minimal inline: update damage_taken log non critico.
-          }
-          turnLogEntry.damage_applied = res.damageDealt;
-          turnLogEntry.roll = res.result.roll;
-          turnLogEntry.hit = res.result.hit;
-          iaActions.push({
-            actor: 'sistema',
-            unit_id: actorId,
-            type: 'attack',
-            target: target.id,
-            die: res.result.die,
-            roll: res.result.roll,
-            mos: res.result.mos,
-            result: res.result.hit ? 'hit' : 'miss',
-            pt: res.result.pt,
-            damage_dealt: res.damageDealt,
-            trait_effects: res.evaluation.trait_effects,
-            actor_position: { ...actor.position },
-            target_position: targetPosAtk,
-            ia_rule: action.source_ia_rule,
-            parry: res.parry,
-          });
-        }
-      } else if (action.type === 'move' && action.move_to) {
-        const dest = action.move_to;
-        const positionFrom = { ...actor.position };
-        // Overlap guard (safety net, gia' fatto in declareSistemaIntents)
-        const blocker = session.units.find(
-          (u) =>
-            u.id !== actor.id && u.hp > 0 && u.position.x === dest.x && u.position.y === dest.y,
-        );
-        if (blocker) {
-          iaActions.push({
-            actor: 'sistema',
-            unit_id: actorId,
-            type: 'skip',
-            reason: `blocked by ${blocker.id} at (${dest.x},${dest.y})`,
-            position_from: positionFrom,
-            position_to: positionFrom,
-            ia_rule: action.source_ia_rule,
-          });
-          turnLogEntry.skipped = 'blocked';
-        } else {
-          actor.position = { x: dest.x, y: dest.y };
-          actor.ap_remaining = Math.max(0, (actor.ap_remaining ?? actor.ap) - 1);
-          const newFacing = facingFromMove(positionFrom, actor.position);
-          if (newFacing) actor.facing = newFacing;
-          const event = buildMoveEvent({ session, actor, positionFrom });
-          event.actor_id = 'sistema';
-          event.ia_rule = action.source_ia_rule;
-          event.ia_controlled_unit = actor.id;
-          session.events.push(event);
-          session.action_counter++;
-          iaActions.push({
-            actor: 'sistema',
-            unit_id: actorId,
-            type: 'move',
-            target: action.target_id || null,
-            position_from: positionFrom,
-            position_to: { ...actor.position },
-            ia_rule: action.source_ia_rule,
-          });
-        }
-      } else {
-        // Altri type: AP only consumption
-        actor.ap_remaining = Math.max(0, (actor.ap_remaining ?? actor.ap) - 1);
-      }
-
-      // Sync back per orchestrator state
-      const uOrch = next.units.find((u) => u.id === actorId);
-      if (uOrch) {
-        if (uOrch.hp) {
-          uOrch.hp.current = Number(actor.hp || 0);
-          uOrch.hp.max = Number(actor.max_hp || actor.hp || 0);
-        }
-        if (uOrch.ap) {
-          uOrch.ap.current = Number(
-            actor.ap_remaining != null ? actor.ap_remaining : actor.ap || 0,
-          );
-        }
-      }
-      // Anche target per attack
-      if (action.type === 'attack' && action.target_id) {
-        const tgtLegacy = session.units.find((u) => u.id === String(action.target_id));
-        const tgtOrch = next.units.find((u) => u.id === String(action.target_id));
-        if (tgtLegacy && tgtOrch && tgtOrch.hp) {
-          tgtOrch.hp.current = Number(tgtLegacy.hp || 0);
-        }
-      }
-      return { nextState: next, turnLogEntry };
-    };
-
-    const result = resolveRoundPure(session.roundState, null, rng, realResolveAction);
-    session.roundState = result.nextState;
-
-    // Persist events written synchronously above
-    await persistEvents(session);
-
-    // 7. Advance turn counter (legacy compat for UI badge)
-    session.turn += 1;
-
-    return {
-      session_id: session.session_id,
-      turn: session.turn,
-      active_unit: session.active_unit,
-      ia_actions: iaActions,
-      ia_action: iaActions[0] || null,
-      side_effects: bleedingEvents,
-      state: publicSessionView(session),
-      round_wrapper: true,
-      round_phase: session.roundState.round_phase,
-      round_decisions: decisions,
-    };
-  }
-
-  router.post('/declare-intent', (req, res, next) => {
-    try {
-      if (!roundModelGuard(req, res)) return;
-      const { session_id: sessionId, actor_id: actorId, action } = req.body || {};
-      const { error, session } = resolveSession(sessionId);
-      if (error) return res.status(error.status).json(error.body);
-      if (!actorId || typeof actorId !== 'string') {
-        return res.status(400).json({ error: 'actor_id richiesto (string)' });
-      }
-      if (!action || typeof action !== 'object') {
-        return res
-          .status(400)
-          .json({ error: "action richiesto (object con campi 'type', 'actor_id', ...)" });
-      }
-      const state = ensureRoundState(session);
-      // beginRound automatico se phase e' null o 'resolved'
-      let current = state;
-      if (current.round_phase !== PHASE_PLANNING) {
-        current = roundOrchestrator.beginRound(current).nextState;
-      }
-      const { nextState } = roundOrchestrator.declareIntent(current, actorId, action);
-      session.roundState = nextState;
-      res.json({
-        session_id: session.session_id,
-        round_phase: nextState.round_phase,
-        pending_intents: nextState.pending_intents,
-      });
-    } catch (err) {
-      // Phase-machine errors -> 400
-      if (err && /round_phase|unit_id/.test(String(err.message || ''))) {
-        return res.status(400).json({ error: err.message });
-      }
-      next(err);
-    }
-  });
-
-  router.post('/clear-intent/:actorId', (req, res, next) => {
-    try {
-      if (!roundModelGuard(req, res)) return;
-      const sessionId = (req.body && req.body.session_id) || req.query.session_id;
-      const actorId = req.params.actorId;
-      const { error, session } = resolveSession(sessionId);
-      if (error) return res.status(error.status).json(error.body);
-      if (!session.roundState) {
-        return res
-          .status(400)
-          .json({ error: 'roundState non inizializzato (chiama prima /declare-intent)' });
-      }
-      const { nextState } = roundOrchestrator.clearIntent(session.roundState, actorId);
-      session.roundState = nextState;
-      res.json({
-        session_id: session.session_id,
-        round_phase: nextState.round_phase,
-        pending_intents: nextState.pending_intents,
-      });
-    } catch (err) {
-      next(err);
-    }
-  });
-
-  router.post('/commit-round', (req, res, next) => {
-    try {
-      if (!roundModelGuard(req, res)) return;
-      const sessionId = req.body && req.body.session_id;
-      const { error, session } = resolveSession(sessionId);
-      if (error) return res.status(error.status).json(error.body);
-      if (!session.roundState) {
-        return res
-          .status(400)
-          .json({ error: 'roundState non inizializzato (chiama prima /declare-intent)' });
-      }
-      const { nextState } = roundOrchestrator.commitRound(session.roundState);
-      session.roundState = nextState;
-      res.json({
-        session_id: session.session_id,
-        round_phase: nextState.round_phase,
-        pending_intents: nextState.pending_intents,
-      });
-    } catch (err) {
-      if (err && /round_phase/.test(String(err.message || ''))) {
-        return res.status(400).json({ error: err.message });
-      }
-      next(err);
-    }
-  });
-
-  router.post('/resolve-round', (req, res, next) => {
-    try {
-      if (!roundModelGuard(req, res)) return;
-      const sessionId = req.body && req.body.session_id;
-      const { error, session } = resolveSession(sessionId);
-      if (error) return res.status(error.status).json(error.body);
-      if (!session.roundState) {
-        return res
-          .status(400)
-          .json({ error: 'roundState non inizializzato (chiama prima /declare-intent)' });
-      }
-      const result = roundOrchestrator.resolveRound(session.roundState);
-      session.roundState = result.nextState;
-      res.json({
-        session_id: session.session_id,
-        round_phase: result.nextState.round_phase,
-        turn_log_entries: result.turnLogEntries,
-        resolution_queue: result.resolutionQueue,
-        reactions_triggered: result.reactionsTriggered,
-        skipped: result.skipped,
-        units: result.nextState.units,
-      });
-    } catch (err) {
-      if (err && /round_phase/.test(String(err.message || ''))) {
-        return res.status(400).json({ error: err.message });
-      }
-      next(err);
-    }
-  });
+  // Round endpoints mounted from sessionRoundBridge.js (token optimization).
+  roundBridge.mountRoundEndpoints(router);
 
   return router;
 }
